SlackSearch/requirements.txt
flask
slack_bolt
google-auth-oauthlib
google-api-python-client
google-oauth2-tool
psycopg2
httplib2